+++ src/entities/auth/sign-up/ui/stages/SecondStage.tsx
@@ -1,0 +1,3 @@
+export default function SecondStage() {
+  return <div></div>;
+}
